+++ unixfs/CHANGELOG.md
@@ -4,6 +4,7 @@
   compatability improvements with go and js equivalents.
 - feat: Add Walker::new_shallow so we can walk directories without loading the entire tree.
 - feat: Support custom hasher.
+- refactor: Change to use CIDv1 by default instead of CIDv0.
 
 # 0.5.1